Extract :jacoco test fixtures
