+++ subprojects/jacoco/jacoco.gradle.kts
@@ -38,9 +38,13 @@ dependencies {
     implementation(library("commons_lang"))
     implementation(library("inject"))
 
+    testFixturesImplementation(project(":baseServices"))
+    testFixturesImplementation(project(":coreApi"))
+    testFixturesImplementation(project(":core"))
+    testFixturesImplementation(project(":internalIntegTesting"))
+    testFixturesImplementation(testLibrary("jsoup"))
+
     testImplementation(project(":fileCollections"))
-    testImplementation(project(":internalIntegTesting"))
-    testImplementation(testLibrary("jsoup"))
     testImplementation(testFixtures(project(":core")))
 
     testRuntimeOnly(project(":runtimeApiInfo"))
